enable assert_trap
