+++ vm/error.go
@@ -21,15 +21,15 @@ func NewExecError(message string) *ExecError {
 var (
 	ErrInvalidBreak           = NewExecError("invalid break recover")
 	ErrTooManyBrTableTarget   = NewExecError("too many br_table targets")
-	ErrIntegerDivisionByZero  = NewExecError("integer division by zero")
+	ErrIntegerDivisionByZero  = NewExecError("integer divide by zero")
 	ErrInvalidIntConversion   = NewExecError("invalid conversion to integer")
 	ErrIntegerOverflow        = NewExecError("integer overflow")
 	ErrInvalidBreakDepth      = NewExecError("invalid break depth")
 	ErrInvalidFunctionBreak   = NewExecError("cannot break out of current function")
-	ErrMismatchedFuncSig      = NewExecError("mismatch function signature")
+	ErrMismatchedFuncSig      = NewExecError("indirect call type mismatch")
 	ErrNoMatchingIfBlock      = NewExecError("no matching If for Else block")
-	ErrOutOfBoundTableAccess  = NewExecError("out of bound table access")
-	ErrOutOfBoundMemoryAccess = NewExecError("out of bound memory access")
+	ErrOutOfBoundTableAccess  = NewExecError("out of bounds table access")
+	ErrOutOfBoundMemoryAccess = NewExecError("out of bounds memory access")
 	ErrUnknownOpcode          = NewExecError("unknown opcode")
 	ErrUnknownReturnType      = NewExecError("unknown block return type")
 	ErrLebOverflow            = NewExecError("unsigned leb overflow")
