+++ vm/vm_test.go
@@ -1,15 +1,11 @@
 package vm
 
 import (
-	"encoding/json"
 	"errors"
 	"fmt"
-	"io"
 	"io/ioutil"
 	"log"
 	"os/exec"
-	"reflect"
-	"strconv"
 	"testing"
 )
 
@@ -49,37 +45,6 @@ type vmTest struct {
 	entry       string
 	expectedErr error
 	trapText    string
-}
-
-func getVM(name string, gasPolicy GasPolicy, gasLimit uint64) *VM {
-	wat := fmt.Sprintf("./test_data/%s.wat", name)
-	wasm := fmt.Sprintf("./test_data/%s.wasm", name)
-	cmd := exec.Command("wat2wasm", wat, "-o", wasm)
-	err := cmd.Start()
-	if err != nil {
-		panic(err)
-	}
-	err = cmd.Wait()
-	if err != nil {
-		panic(err)
-	}
-	data, err := ioutil.ReadFile(wasm)
-	if err != nil {
-		panic(err)
-	}
-	vm, err := NewVM(data, gasPolicy, &Gas{Limit: gasLimit}, &TestResolver{})
-	if err != nil {
-		panic(err)
-	}
-	return vm
-}
-
-func TestNeg(t *testing.T) {
-	vm := getVM("i32", &FreeGasPolicy{}, 0)
-	_, ok := vm.GetFunctionIndex("somefunc")
-	if ok {
-		t.Errorf("Expect function index to be -1")
-	}
 }
 
 type TestResolver struct{}
@@ -135,7 +100,38 @@ func (r *TestResolver) GetFunction(module, name string) HostFunction {
 	return nil
 }
 
-func TestVmError(t *testing.T) {
+func GetTestVM(name string, gasPolicy GasPolicy, gasLimit uint64) *VM {
+	wat := fmt.Sprintf("./test_data/%s.wat", name)
+	wasm := fmt.Sprintf("./test_data/%s.wasm", name)
+	cmd := exec.Command("wat2wasm", wat, "-o", wasm)
+	err := cmd.Start()
+	if err != nil {
+		panic(err)
+	}
+	err = cmd.Wait()
+	if err != nil {
+		panic(err)
+	}
+	data, err := ioutil.ReadFile(wasm)
+	if err != nil {
+		panic(err)
+	}
+	vm, err := NewVM(data, gasPolicy, &Gas{Limit: gasLimit}, &TestResolver{})
+	if err != nil {
+		panic(err)
+	}
+	return vm
+}
+
+func TestGetFunctionIndex(t *testing.T) {
+	vm := GetTestVM("i32", &FreeGasPolicy{}, 0)
+	_, ok := vm.GetFunctionIndex("somefunc")
+	if ok {
+		t.Errorf("Expect function index to be -1")
+	}
+}
+
+func TestVMError(t *testing.T) {
 	tests := []vmTest{
 		{name: "exit", entry: "calc", params: []uint64{1}, expectedErr: ErrUnreachable},
 		{name: "local", entry: "calc", params: []uint64{}, expectedErr: ErrWrongNumberOfArgs},
@@ -163,7 +159,7 @@ func TestVmError(t *testing.T) {
 				}
 			}()
 
-			vm := getVM(test.name, &FreeGasPolicy{}, 0)
+			vm := GetTestVM(test.name, &FreeGasPolicy{}, 0)
 			fnID, ok := vm.GetFunctionIndex(test.entry)
 			if !ok {
 				t.Error("cannot get function export")
@@ -204,8 +200,7 @@ func TestVM(t *testing.T) {
 		{name: "trunc_edge", entry: "main", params: []uint64{}, expected: 0},
 	}
 	for _, test := range tests {
-		vm := getVM(test.name, &FreeGasPolicy{}, 0)
-		fmt.Println(vm.Module.TableIndexSpace[0])
+		vm := GetTestVM(test.name, &FreeGasPolicy{}, 0)
 
 		fnID, ok := vm.GetFunctionIndex(test.entry)
 		if !ok {
@@ -224,176 +219,8 @@ func TestVM(t *testing.T) {
 	}
 }
 
-func TestWasmSuite(t *testing.T) {
-	tests := []string{
-		"i32", "i64", "f32", "f64",
-		"f32_cmp", "f32_bitwise", "f64_cmp", "f64_bitwise",
-		"br", "br_if", "br_table",
-		"call", "call_indirect",
-		"global", "local_get", "local_set", "local_tee",
-		"memory", "memory_grow", "memory_size", "memory_redundancy", "memory_trap",
-		"binary", "binary-leb128", "block",
-		"address",
-		"comments",
-		"return", "select", "loop", "if",
-		"custom", "endianness",
-		"fac", "float_literals", "float_memory",
-		"forward", "func",
-		"inline-module", "int_exprs", "int_literals", "labels",
-		"left-to-right", "load", "nop", "stack", "store", "switch", "token",
-		"traps", "type", "unreachable", "unreached-invalid", "unwind",
-		"utf8-custom-section-id", "utf8-import-field", "utf8-import-module", "utf8-invalid-encoding",
-		"skip-stack-guard-page", "float_exprs", "float_misc", "align",
-		"start", "func_ptrs",
-		"const", "table", "break-drop",
-		"conversions",
-
-		// "exports", // empty module removed
-		// "linking",
-		// "elem", "data", //wagon parsing failed
-		// "names", // problem with unicode. Entries key and cmd.Action.Field yield different codes
-		// "imports", // missing imports from spec
-	}
-
-	for _, name := range tests {
-		t.Logf("Test suite %s", name)
-		wast := fmt.Sprintf("./test_suite/%s.wast", name)
-		jsonFile := fmt.Sprintf("./test_suite/%s.json", name)
-		cmd := exec.Command("wast2json", wast, "-o", jsonFile)
-		err := cmd.Start()
-		if err != nil {
-			panic(err)
-		}
-		err = cmd.Wait()
-		if err != nil {
-			panic(err)
-		}
-
-		raw, err := ioutil.ReadFile(jsonFile)
-		if err != nil {
-			panic(err)
-		}
-		var suite TestSuite
-		err = json.Unmarshal(raw, &suite)
-		if err != nil {
-			panic(err)
-		}
-		var vm *VM
-		for _, cmd := range suite.Commands {
-			t.Logf("Running test %s %d", name, cmd.Line)
-			if cmd.Action.Field == "as-unary-operand" && cmd.Line == 338 {
-				continue
-			}
-			if name == "linking" && cmd.Line >= 50 && cmd.Line <= 83 {
-				continue
-			}
-			// Skip min, max nan with inf tests
-			if (name == "f32" || name == "f64") && ((cmd.Line >= 1931 && cmd.Line <= 1938) ||
-				(cmd.Line >= 1995 && cmd.Line <= 2002) ||
-				(cmd.Line >= 2331 && cmd.Line <= 2338) ||
-				(cmd.Line >= 2395 && cmd.Line <= 2402)) {
-				continue
-			}
-			switch cmd.Type {
-			case "module":
-				data, err := ioutil.ReadFile(fmt.Sprintf("./test_suite/%s", cmd.Filename))
-				if err != nil {
-					t.Error(err)
-				}
-				vm, err = NewVM(data, &FreeGasPolicy{}, &Gas{}, &TestResolver{})
-				if err != nil {
-					t.Error(err)
-				}
-			case "assert_return", "action", "assert_return_canonical_nan", "assert_return_arithmetic_nan":
-				switch cmd.Action.Type {
-				case "invoke":
-					funcID, ok := vm.GetFunctionIndex(cmd.Action.Field)
-					if !ok {
-						t.Errorf("function not found %s", cmd.Action.Field)
-						continue
-					}
-					args := make([]uint64, 0)
-					for _, arg := range cmd.Action.Args {
-						val, err := strconv.ParseUint(arg.Value, 10, 64)
-						if err != nil {
-							panic(err)
-						}
-						args = append(args, val)
-					}
-					// t.Logf("Triggering %s with args at line %d", cmd.Action.Field, cmd.Line)
-					// t.Log(args)
-					ret, _ := vm.Invoke(funcID, args...)
-					// t.Log("ret", ret)
-
-					if len(cmd.Expected) != 0 {
-						var exp uint64
-						// t.Log("cmd.Type", cmd.Type, cmd.Expected[0].Value)
-						if cmd.Expected[0].Value == "nan:canonical" {
-							if cmd.Expected[0].Type == "f32" {
-								exp = 0x7fc00000
-							} else if cmd.Expected[0].Type == "f64" {
-								exp = 0x7ff8000000000000
-							}
-						} else if cmd.Expected[0].Value == "nan:arithmetic" {
-							// An arithmetic NaN is a floating-point value ±𝗇𝖺𝗇(n) with n≥canonN, such that the most significant bit is 1 while all others are arbitrary.
-							// Unset sign bit, pass if >= canonical NaN in integer
-							if cmd.Expected[0].Type == "f32" && (uint32(ret)&^(1<<31)) >= uint32(0x7fc00000) {
-								exp = ret
-							}
-							if cmd.Expected[0].Type == "f64" && (ret&^(1<<63)) >= uint64(0x7ff8000000000000) {
-								exp = ret
-							}
-						} else {
-							exp, err = strconv.ParseUint(cmd.Expected[0].Value, 10, 64)
-							if err != nil {
-								panic(err)
-							}
-						}
-
-						if cmd.Expected[0].Type == "i32" || cmd.Expected[0].Type == "f32" {
-							ret = uint64(uint32(ret))
-							exp = uint64(uint32(exp))
-						}
-
-						if ret != exp {
-							t.Errorf("Test %s Field %s Line %d: Expect return value to be %d, got %d", name, cmd.Action.Field, cmd.Line, exp, ret)
-						}
-					}
-				case "get":
-					entry, ok := vm.Module.ExportSec.ExportMap[cmd.Action.Field]
-					if !ok {
-						panic("Global export not found")
-					}
-					ret := vm.globals[entry.Desc.Idx]
-					if len(cmd.Expected) != 0 {
-						exp, err := strconv.ParseUint(cmd.Expected[0].Value, 10, 64)
-						if err != nil {
-							panic(err)
-						}
-
-						if cmd.Expected[0].Type == "i32" || cmd.Expected[0].Type == "f32" {
-							ret = uint64(uint32(ret))
-							exp = uint64(uint32(exp))
-						}
-						if ret != exp {
-							t.Errorf("Test %s Field %s Line %d: Expect return value to be %d, got %d", name, cmd.Action.Field, cmd.Line, exp, ret)
-						}
-					}
-				default:
-					t.Errorf("unknown action %s", cmd.Action.Type)
-				}
-
-			case "assert_trap", "assert_invalid", "assert_exhaustion", "assert_malformed", "assert_uninstantiable", "assert_unlinkable":
-				t.Logf("Skipping %s", cmd.Type)
-			default:
-				t.Errorf("unknown command %s", cmd.Type)
-			}
-		}
-	}
-}
-
 func TestEnoughGas(t *testing.T) {
-	vm := getVM("i32", &SimpleGasPolicy{}, 2148)
+	vm := GetTestVM("i32", &SimpleGasPolicy{}, 2148)
 	fnIndex, ok := vm.GetFunctionIndex("calc")
 	if !ok {
 		panic("Cannot get export fn index")
@@ -405,7 +232,7 @@ func TestEnoughGas(t *testing.T) {
 }
 
 func TestOutOfGas(t *testing.T) {
-	vm := getVM("i32", &SimpleGasPolicy{}, 2058)
+	vm := GetTestVM("i32", &SimpleGasPolicy{}, 2058)
 	fnIndex, ok := vm.GetFunctionIndex("calc")
 	if !ok {
 		panic("Cannot get export fn index")
@@ -416,113 +243,8 @@ func TestOutOfGas(t *testing.T) {
 	}
 }
 
-func TestMemSize(t *testing.T) {
-	vm := getVM("i32", &FreeGasPolicy{}, 0)
-	if len(vm.memory) != vm.MemSize() {
-		t.Errorf("Expect MemSize to be %d, got %d", len(vm.memory), vm.MemSize())
-	}
-}
-
-func TestMemGrow(t *testing.T) {
-	vm := getVM("memory_grow", &SimpleGasPolicy{}, 1024*3+3)
-	fnIndex, ok := vm.GetFunctionIndex("grow")
-	if !ok {
-		panic("Cannot get export fn index")
-	}
-	_, err := vm.Invoke(fnIndex)
-	if err != nil {
-		t.Errorf("Expect execution to go through, got %v", err)
-	}
-}
-
-func TestMemGrowOutOfGas(t *testing.T) {
-	vm := getVM("memory_grow", &SimpleGasPolicy{}, 1024*2+3)
-	fnIndex, ok := vm.GetFunctionIndex("grow")
-	if !ok {
-		panic("Cannot get export fn index")
-	}
-	_, err := vm.Invoke(fnIndex)
-	if err != ErrOutOfGas {
-		t.Errorf("Expect execution to be out of gas, got %v", err)
-	}
-}
-
-func TestMemInitOutOfGas(t *testing.T) {
-	defer func() {
-		if r := recover(); r != nil {
-			err := r.(error)
-			if err != ErrOutOfGas {
-				t.Errorf("Expect execution to be out of gas, got %v", err)
-			}
-		}
-	}()
-	getVM("memory_grow", &SimpleGasPolicy{}, 2047)
-}
-
-func TestMemRead(t *testing.T) {
-	vm := getVM("i32", &FreeGasPolicy{}, 0)
-	sample := []byte{0, 1, 2, 3, 4, 5, 6, 7, 8, 9}
-	offset := vm.MemSize() - len(sample)
-	copy(vm.memory[offset:offset+len(sample)], sample)
-	readBuffer := make([]byte, 10)
-	readSize, err := vm.MemRead(readBuffer, offset)
-	if readSize != len(sample) {
-		t.Errorf("Expect MemRead result size to be %d, got %d", len(sample), readSize)
-	}
-	if err != nil {
-		t.Errorf("Expect MemRead err to be nil, got %d", err)
-	}
-	if !reflect.DeepEqual(sample, readBuffer) {
-		t.Errorf("Expect MemRead result to be %v, got %v", sample, readBuffer)
-	}
-
-	readBuffer = make([]byte, 15)
-	readSize, err = vm.MemRead(readBuffer, offset)
-	if readSize != len(sample) {
-		t.Errorf("Expect MemRead result size to be %d, got %d", len(sample), readSize)
-	}
-	if err != io.ErrShortBuffer {
-		t.Errorf("Expect MemRead err to be io.ErrShortBuffer, got %v", err)
-	}
-	if !reflect.DeepEqual(sample, readBuffer[:len(sample)]) {
-		t.Errorf("Expect MemRead result first 10 bytes to be %v, got %v", sample, readBuffer)
-	}
-
-}
-
-func TestMemWrite(t *testing.T) {
-	vm := getVM("i32", &FreeGasPolicy{}, 0)
-	sample := []byte{0, 1, 2, 3, 4, 5, 6, 7, 8, 9}
-	offset := vm.MemSize() - len(sample)
-	writeSize, err := vm.MemWrite(sample, offset)
-
-	if writeSize != len(sample) {
-		t.Errorf("Expect MemWrite result size to be %d, got %d", len(sample), writeSize)
-	}
-	if err != nil {
-		t.Errorf("Expect MemWrite err to be nil, got %d", err)
-	}
-	if !reflect.DeepEqual(sample, vm.memory[offset:offset+len(sample)]) {
-		t.Errorf("Expect MemWrite result to be %v, got %v", sample, vm.memory[offset:offset+len(sample)])
-	}
-
-	sample = []byte{0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12}
-
-	writeSize, err = vm.MemWrite(sample, offset)
-
-	if writeSize != vm.MemSize()-offset {
-		t.Errorf("Expect MemWrite result size to be %d, got %d", vm.MemSize()-offset, writeSize)
-	}
-	if err != io.ErrShortWrite {
-		t.Errorf("Expect MemWrite err to be io.ErrShortWrite, got %d", err)
-	}
-	if !reflect.DeepEqual(sample[:writeSize], vm.memory[offset:]) {
-		t.Errorf("Expect MemWrite result to be %v, got %v", sample[:writeSize], vm.memory[offset:])
-	}
-}
-
 func TestOutOfGasVMCreation(t *testing.T) {
-	getVM("i32", &FreeGasPolicy{}, 0)
+	GetTestVM("i32", &FreeGasPolicy{}, 0)
 	data, err := ioutil.ReadFile("./test_data/i32.wasm")
 	if err != nil {
 		panic(err)
